Move HtmlProcessor into a separate module
