+++ app/core/src/main/java/com/fsck/k9/message/html/HtmlProcessorFactory.kt
@@ -1,5 +1,7 @@
 package com.fsck.k9.message.html
 
+import app.k9mail.html.cleaner.HtmlProcessor
+
 class HtmlProcessorFactory(
     private val displayHtmlFactory: DisplayHtmlFactory
 ) {
