+++ app/html-cleaner/src/main/java/app/k9mail/html/cleaner/BodyCleaner.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.message.html
+package app.k9mail.html.cleaner
 
 import org.jsoup.nodes.Document
 import org.jsoup.safety.Cleaner
