+++ app/html-cleaner/src/main/java/app/k9mail/html/cleaner/HtmlHeadProvider.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.message.html
+package app.k9mail.html.cleaner
 
 interface HtmlHeadProvider {
     val headHtml: String
