+++ app/html-cleaner/src/main/java/app/k9mail/html/cleaner/HtmlSanitizer.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.message.html
+package app.k9mail.html.cleaner
 
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
