+++ build.gradle
@@ -40,6 +40,7 @@ buildscript {
                 'okhttp': '4.9.2',
                 'minidns': '1.0.0',
                 'glide': '4.12.0',
+                'jsoup': '1.13.1',
 
                 'androidxTestRunner': '1.4.0',
                 'junit': '4.13.2',
