+++ settings.gradle
@@ -10,6 +10,7 @@ include ':app:autodiscovery:api'
 include ':app:autodiscovery:providersxml'
 include ':app:autodiscovery:srvrecords'
 include ':app:autodiscovery:thunderbird'
+include ':app:html-cleaner'
 include ':mail:common'
 include ':mail:testing'
 include ':mail:protocols:imap'
